# init-project/evals/fixtures/springboot-commerce-webflux/src/main/java/com/acme/commerce/endpoint/ShipmentEndpoint.java
package com.acme.commerce.endpoint;

import com.acme.commerce.model.CommerceModels.CreateShipmentRequest;
import com.acme.commerce.model.CommerceModels.ShipmentView;
import com.acme.commerce.service.ShipmentService;
import javax.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;
import reactor.core.publisher.Mono;

@RestController
@RequestMapping("/api/v1/shipments")
public class ShipmentEndpoint {
    private final ShipmentService shipmentService;

    public ShipmentEndpoint(ShipmentService shipmentService) {
        this.shipmentService = shipmentService;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public Mono<ShipmentView> createShipment(@Valid @RequestBody CreateShipmentRequest request) {
        return shipmentService.createShipment(request);
    }

    @GetMapping("/{shipmentId}")
    public Mono<ShipmentView> getShipment(@PathVariable String shipmentId) {
        return shipmentService.getShipment(shipmentId);
    }
}